Refactored the translation service
