+++ src/Magister/Config/dictionary.php
@@ -1,0 +1,24 @@
+<?php
+
+return [
+	/*
+     |--------------------------------------------------------------------------
+     | General
+     |--------------------------------------------------------------------------
+     */
+    'href' => 'Href',
+	/*
+     |--------------------------------------------------------------------------
+     | Grade
+     |--------------------------------------------------------------------------
+     */
+    'mark' => 'CijferStr',
+
+   	/*
+     |--------------------------------------------------------------------------
+     | Appointment
+     |--------------------------------------------------------------------------
+     */
+    'description' => 'Omschrijving',
+    
+];
