+++ src/Magister/Services/Contracts/Translation/ShouldBeTranslatable.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Magister\Services\Contracts\Translation;
+
+interface ShouldBeTranslatable
+{
+    //
+}
