+++ src/Magister/Services/Support/helpers.php
@@ -10,11 +10,7 @@ if (!function_exists('trans')) {
     {
         $translator = \App::make('Magister\Services\Translation\Translator');
         
-        if ($translator->translationExistsFor($foreign)) {
-            return $translator->translateForeign($foreign);
-        }
-
-        return $foreign;
+        return $translator->translateForeign($foreign);
     }
 }
