+++ src/Magister/Services/Translation/TranslationNotFoundException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Magister\Services\Translation;
+
+use Exception;
+
+class TranslationNotFoundException extends Exception
+{
+	//
+}
